Change where custom songs are intercepted. Move songloader to class
