+++ Patches/MusicManager_Awake.cs
@@ -11,6 +11,7 @@ using NickCustomMusicMod.Management;
 using System.Collections;
 using System.IO;
 using UnityEngine.Networking;
+using NASB2CustomMusicMod.Utils;
 
 namespace NASB2CustomMusicMod.Patches
 {
@@ -19,21 +20,6 @@ namespace NASB2CustomMusicMod.Patches
     {
         static bool Prefix(MusicManager __instance)
         {
-            for (int i = 0; i < __instance.MusicSources.Length; ++i) {
-                AudioSource musicSource = __instance.MusicSources[i];
-
-                string id = musicSource.clip.name;
-                //Plugin.LogInfo($"MusicManager: {__instance.name} | MusicSources[{i}]: {id}");
-
-                //if (CheckToSkipOnlineMenuMusic(ref id)) return false;
-
-                MusicTrack customTrack = CustomMusicManager.GetRandomCustomSong(id);
-
-                if (customTrack == null) continue;
-
-                __instance.StartCoroutine(LoadCustomSong(customTrack, musicSource));
-            }
-
             return true;
         }
 
@@ -62,50 +48,6 @@ namespace NASB2CustomMusicMod.Patches
             Plugin.previousMusicID = id;
 
             return false;
-        }
-
-        public static IEnumerator LoadCustomSong(MusicTrack track, AudioSource musicSource)
-        {
-            //Plugin.LogInfo("Loading song: " + track.Path);
-
-            UnityEngine.AudioType audioType = UnityEngine.AudioType.UNKNOWN;
-
-            switch (Path.GetExtension(track.Path).ToLower())
-            {
-                case ".wav":
-                    audioType = UnityEngine.AudioType.WAV;
-                    break;
-                case ".mp3":
-                    audioType = UnityEngine.AudioType.MPEG;
-                    break;
-                case ".ogg":
-                    audioType = UnityEngine.AudioType.OGGVORBIS;
-                    break;
-                default:
-                    yield break;
-            }
-
-            UnityWebRequest audioLoader = UnityWebRequestMultimedia.GetAudioClip(track.Path, audioType);
-            // this stops the lag!
-            (audioLoader.downloadHandler as DownloadHandlerAudioClip).streamAudio = true;
-            yield return audioLoader.SendWebRequest();
-            if (audioLoader.error != null)
-            {
-                Plugin.LogError(audioLoader.error);
-                yield break;
-            }
-
-            musicSource.clip = DownloadHandlerAudioClip.GetContent(audioLoader);
-            //musicSource.volume = 1;
-
-            // !!!!!! TODO !!!!!!
-            // !!!!!! TODO !!!!!!
-            // !!!!!! TODO !!!!!!
-            // !!!!!! TODO !!!!!!
-            // Handle custom music data here
-            // HandleCustomMusicData(entry, music);
-
-            yield break;
         }
     }
 }
